requirements.txt
django
drf-spectacular
djangorestframework==3.14.0
flake8>=6.0.0,<6.1.0
black>=23.3.0,<23.4.0
mysqlclient
